requirements.txt
Flask==2.0.1
fastapi==0.68.0
uvicorn==0.15.0
web3==5.24.0
pycryptodome==3.10.1
pandas==1.3.3
openpyxl==3.0.9
requests==2.26.0
python-dotenv==0.19.1